+++ flutter_front/lib/reservation_status/view/reservation_status_view.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_front/common/component/container/responsive_container.dart';
 import 'package:flutter_front/common/component/container/stack_container.dart';
-import 'package:flutter_front/common/component/custome_icon_button.dart';
+import 'package:flutter_front/common/component/custom_icon_button.dart';
 import 'package:flutter_front/common/styles/sizes.dart';
 import 'package:flutter_front/common/styles/text_styles.dart';
 import 'package:flutter_front/common/utils/date_utils.dart';
