+++ flutter_front/lib/user_management/view/user_info_container.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_front/common/component/container/designed_container.dart';
 import 'package:flutter_front/common/component/container/responsive_container.dart';
-import 'package:flutter_front/common/component/custome_icon_button.dart';
+import 'package:flutter_front/common/component/custom_icon_button.dart';
 import 'package:flutter_front/common/component/custome_titled_text.dart';
 import 'package:flutter_front/common/styles/sizes.dart';
 import 'package:flutter_front/common/styles/text_styles.dart';
